CBL-8192 : Update to handle build errors when using Xcode 26.4 (#3514)
Fix build errors for the followings:

- Implicit conversion from string to ‘double' may lose preciscion

- ISO C++ requires field designators to be specified in declaration order

+++ Objective-C/CBLCollection.mm
@@ -455,7 +455,7 @@ NSString* const kCBLDefaultCollectionName = @"_default";
         if (timestamp == 0) {
             return nil;
         }
-        return [NSDate dateWithTimeIntervalSince1970: (timestamp/msec)];
+        return [NSDate dateWithTimeIntervalSince1970: ((double)timestamp/msec)];
     }
 }
